# packages/contracts/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/companies": {
      "post": {
        "description": "支持 Idempotency-Key 头：同 key 重放返回首次结果，不重复创建（PRD 11.16）。",
        "operationId": "CompanyController_create_v1",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": false,
            "in": "header",
            "description": "幂等键（客户端生成，如 uuid）",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CompanyDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "提交官网，创建企业画像并触发理解（异步）",
        "tags": [
          "Companies"
        ],
        "x-required-scopes": [
          "acquisition:write"
        ]
      },
      "get": {
        "operationId": "CompanyController_list_v1",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "page"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CompanyDto"
                      }
                    },
                    "page": {
                      "type": "object",
                      "required": [
                        "next_cursor",
                        "has_more"
                      ],
                      "properties": {
                        "next_cursor": {
                          "type": "string",
                          "nullable": true,
                          "description": "下一页游标；null = 没有更多"
                        },
                        "has_more": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "列出当前 workspace 的企业画像（游标分页）",
        "tags": [
          "Companies"
        ],
        "x-required-scopes": [
          "acquisition:read"
        ]
      }
    },
    "/api/v1/companies/{companyId}": {
      "get": {
        "operationId": "CompanyController_get_v1",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CompanyDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "按 id 获取企业画像（跨租户返回 404）",
        "tags": [
          "Companies"
        ],
        "x-required-scopes": [
          "acquisition:read"
        ]
      }
    },
    "/api/v1/companies/{companyId}/completeness": {
      "get": {
        "operationId": "CompanyController_completeness_v1",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": [
                        "status",
                        "approvedClaims",
                        "pendingClaims",
                        "offerings",
                        "conflictsOpen"
                      ],
                      "properties": {
                        "status": {
                          "type": "string"
                        },
                        "approvedClaims": {
                          "type": "integer"
                        },
                        "pendingClaims": {
                          "type": "integer"
                        },
                        "offerings": {
                          "type": "integer"
                        },
                        "conflictsOpen": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "企业完整度（5.2.7）：审批数/待审数/产品数/未决冲突 + 当前状态",
        "tags": [
          "Companies"
        ],
        "x-required-scopes": [
          "acquisition:read"
        ]
      }
    },
    "/api/v1/companies/{companyId}/confirm": {
      "post": {
        "operationId": "CompanyController_confirm_v1",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CompanyDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "人工确认企业可用（REVIEW → ACTIVE，显式 Gate 出口）",
        "tags": [
          "Companies"
        ],
        "x-required-scopes": [
          "acquisition:review"
        ]
      }
    },
    "/api/v1/companies/{companyId}/offerings": {
      "get": {
        "operationId": "CompanyController_listOfferings_v1",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/OfferingDto"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "企业的结构化产品/服务（理解工作流抽取，逐条带来源页与原文片段）",
        "tags": [
          "Companies"
        ],
        "x-required-scopes": [
          "acquisition:read"
        ]
      }
    },
...
  },
  "info": {
    "title": "Global API",
    "description": "出海企业 AI 全球客户开发与增长平台 · 后端 API（前端接入见 packages/contracts/INTEGRATION.md）",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "同源部署（相对路径；具体 host 由部署环境决定）"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "PublicContactDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "email",
              "phone",
              "social"
            ]
          },
          "value": {
            "type": "string",
            "example": "sales@acme-tech.com"
          },
          "sourceUrl": {
            "type": "string",
            "description": "该联系方式被发现的页面 URL（溯源）"
          }
        },
        "required": [
          "type",
          "value",
          "sourceUrl"
        ]
      },
      "CompanyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Shenzhen Acme Tech Co."
          },
          "website": {
            "type": "string",
            "nullable": true,
            "example": "https://acme-tech.com"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ENRICHING",
              "REVIEW",
              "ACTIVE"
            ]
          },
          "industry": {
            "type": "string",
            "nullable": true,
            "description": "行业（理解工作流回填）"
          },
          "summary": {
            "type": "string",
            "nullable": true,
            "description": "企业简介（抽取，可人工修正）"
          },
          "publicContacts": {
            "description": "官网确定性抽取的公开联系方式/社媒（正则命中，非 AI 生成，逐条带来源页）",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicContactDto"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "website",
          "status",
          "industry",
          "summary",
          "publicContacts",
          "createdAt"
        ]
      },
      "CreateCompanyDto": {
        "type": "object",
        "properties": {
          "website": {
            "type": "string",
            "example": "https://acme-tech.com",
            "description": "企业官网 URL"
          },
          "name": {
            "type": "string",
            "example": "Shenzhen Acme Tech Co.",
            "description": "企业名称（可选；未填则先用域名占位，理解后回填）"
          }
        },
        "required": [
          "website"
        ]
      },
      "OfferingDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "CNC 精密加工件"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "attributes": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true,
            "description": "结构化属性（moq/lead_time/materials/params/certifications…，仅来源文本明确出现的）"
          },
          "sourceUrl": {
            "type": "string",
            "nullable": true,
            "description": "溯源：来自哪个页面"
          },
          "evidence": {
            "type": "string",
            "nullable": true,
            "description": "来源页原文片段"
          },
          "confidence": {
            "type": "number",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "attributes",
          "sourceUrl",
          "evidence",
          "confidence",
          "createdAt"
        ]
      },
...
    }
  }
}